requirements.txt
astroid==2.0.4
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-registration-redux==2.6
django-tastypie==0.14.2
djangorestframework==3.10.3
dodgy==0.1.9
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauthlib==3.1.0
pep8-naming==0.4.1
prospector==1.1.6.2
pycodestyle==2.4.0
pydocstyle==4.0.1
pyflakes==1.6.0
pylint==2.1.1
pylint-celery==0.3
pylint-django==2.0.2
pylint-flask==0.5
pylint-plugin-utils==0.6
python-dateutil==2.8.0
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
requirements-detector==0.6
setoptconf==0.2.0
six==1.12.0
snowballstemmer==1.9.1
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
wrapt==1.11.2
